moved PtNDArray interface out
