+++ extensions/fasttext/src/main/java/ai/djl/fasttext/jni/FtWrapper.java
@@ -13,14 +13,14 @@
 package ai.djl.fasttext.jni;
 
 import ai.djl.modality.Classifications;
-import ai.djl.util.NativeResource;
+import ai.djl.util.NativeResourceImpl;
 
 import java.util.ArrayList;
 import java.util.List;
 
 /** A class containing utilities to interact with the fastText JNI layer. */
 @SuppressWarnings("MissingJavadocMethod")
-public final class FtWrapper extends NativeResource<Long> {
+public final class FtWrapper extends NativeResourceImpl<Long> {
 
     private static RuntimeException libraryStatus;
